Add PATCH /todos

+++ server/server.js
@@ -1,6 +1,7 @@
-var express = require('express');
-var bodyParser = require('body-parser');
-var { mongoose } = require('./db/mongoose.js');
+const  express = require('express');
+const _=require('lodash');
+const  bodyParser = require('body-parser');
+const  { mongoose } = require('./db/mongoose.js');
 const { ObjectID } = require('mongodb');
 
 var { Todo } = require('./models/todo.js');
@@ -47,16 +48,41 @@ app.get('/todos/:id', (req, res) => {
 
 app.delete('/todos/:id',(req,res)=>{
        var id = req.params.id;
-    // if (!ObjectID.isValid(id)) {
-    //     return res.status(404).send();
-    // };
+    if (!ObjectID.isValid(id)) {
+        return res.status(404).send();
+    };
     Todo.findByIdAndRemove(id).then((todo)=>{
-       //     if(!todo){
-       // 	return res.status(404).send();
-       // }
-       res.send(todo);
+           if(!todo){
+       	return res.status(404).send();
+       }
+       res.send({todo:todo});
     }).catch((e) => {
         res.status(400).send();
+    });
+
+});
+
+app.patch('/todos/:id',(req,res)=>{
+  var id=req.params.id;
+  var body=_.pick(req.body,['text','completed']);
+      if (!ObjectID.isValid(id)) {
+        return res.status(404).send();
+    };
+
+    if(_.isBoolean(body.completed)&&body.completed){
+     body.completedAt=new Date().getTime();
+    }
+    else{
+     body.completed=false;
+     body.completedAt=null;
+    }
+    Todo.findByIdAndUpdate(id,{$set:body},{new:true}).then((todo)=>{
+     if(!todo){
+      return res.status(404).send();
+     }
+     res.send({todo});
+    }).catch((e)=>{
+      res.status(400).send();
     });
 
 });
